Fix @Notnull annotation and disable the test which depends on ZK by default.

+++ venice-common/src/test/java/com/linkedin/venice/helix/TestHelixMetadataRepository.java
@@ -18,17 +18,28 @@ public class TestHelixMetadataRepository {
     private ZkClient zkClient;
     private String clusterPath = "/test-metadata-cluster";
     private String storesPath = "/stores";
+    /**
+     * By default, this test is inactive. Because it depends on external zk process. It should be only used in
+     * debugging.
+     */
+    private boolean isActive = false;
 
     @BeforeTest
     public void zkSetup() {
+        if (!isActive) {
+            return;
+        }
         zkClient = new ZkClient(zkAddress, ZkClient.DEFAULT_SESSION_TIMEOUT, ZkClient.DEFAULT_CONNECTION_TIMEOUT,
             new HelixStoreSerializer(new StoreJSONSerializer()));
         zkClient.create(clusterPath, null, CreateMode.PERSISTENT);
-        zkClient.create(clusterPath+storesPath, null, CreateMode.PERSISTENT);
+        zkClient.create(clusterPath + storesPath, null, CreateMode.PERSISTENT);
     }
 
     @AfterTest
     public void zkCleanup() {
+        if (!isActive) {
+            return;
+        }
         zkClient.deleteRecursive(clusterPath);
         zkClient.delete(clusterPath);
         zkClient.close();
@@ -36,7 +47,10 @@ public class TestHelixMetadataRepository {
 
     @Test
     public void testAddAndReadStore() {
-        HelixMetadataRepository repo = new HelixMetadataRepository(zkClient, clusterPath+storesPath);
+        if (!isActive) {
+            return;
+        }
+        HelixMetadataRepository repo = new HelixMetadataRepository(zkClient, clusterPath + storesPath);
         Store s1 = new Store("s1", "owner", System.currentTimeMillis());
         repo.addStore(s1);
         Store s2 = repo.getStore("s1");
@@ -45,7 +59,10 @@ public class TestHelixMetadataRepository {
 
     @Test
     public void testAddAndDeleteStore() {
-        HelixMetadataRepository repo = new HelixMetadataRepository(zkClient, clusterPath+storesPath);
+        if (!isActive) {
+            return;
+        }
+        HelixMetadataRepository repo = new HelixMetadataRepository(zkClient, clusterPath + storesPath);
         Store s1 = new Store("s1", "owner", System.currentTimeMillis());
         repo.addStore(s1);
         repo.deleteStore("s1");
